chore: alert 훅 분리
